Implemented User validation
User now has a common function for making the parameter needed to
validate the User. This is used when sending validation email, and by
UserValidate snippet.
UserValidate snippet now has a companion object with a menu function
that provides a Menu.param handling the actual validation, should be
added to site menu, probably Hidden.
Boot has the user validation menu item as described.