AjaxView for redirecting, UserSignup redirects on completion
AjaxRedirectView will cause the browser to redirect to an url, when a
Ref[Option[String]] changes to contain Some(url). This can be used for
example to redirect on successful completion of a form.
UserSignup now redirects to user_signup_complete when user has been
created and validation mail sent.
Added user_signup_complete.html and corresponding menu entry in Boot.