Fixed handling of Long values, removed support for Floats
MongoTokens had a bug in handling of Long values (and Floats), since it
did not correctly encode them as "second class primitives" (objects
used to represent primitive types that have no equivalent in mongo).
However it turns out that Long values CAN be unambiguously encoded in
mongo, as "NumberLong". This allows us to handle them without the
concept of "second class primitives". Support for Floats has been
dropped for now, since it makes more sens to use Doubles instead, for
Nodes that will be persisted to mongo.