Added partial support for extended sessions
System is loosely based on the Lift extended sessions for mapper.
Currently works with the exception of an annoying bug that when
automatic login is performed, the cookie deletion and recreation does
not seem to make it to the browser, leaving it with an old cookie. This
could be worked around by not renewing the cookie on every automatic
login, but this would be less secure. Note that the same cookie system
seems to work fine for Lift's built-in extended sessions, and S.notice
calls at the same time as the S cookie calls succeed.