Updated to bootstrap 3.0.3 and font-awesome 4.0.3
Requires a surprisingly huge amount of changes to classes etc.
However the html is a little neater now, and we've dumped a huge amount
of stuff in old boxes.css, replaced with much less extra css in the
bootstrap-boxes directory, split into multiple files by area (can be
recombined later).