Added workaround to make ExtendedSession work
Lift currently calls earlyInStateful twice when creating a session as
it serves a page. This happens when the server has been restarted, or a
new browser is used, or the session cookie is lost, etc. This
previously broke the ExtendedSession login handling, since the first
call to earlyInStateful handled logging in, but Lift lost the cookies
it produced. We now use an additional SessionVar to detect the first
earlyInStateful call and ignore it, as suggested by Antonio Salazar
Cardozo :)

ExtendedSessions now work fine, the S.notice and logger calls have been
removed.