Work on angular animations, two-way sync of angular and boxes
Angular animations tweaked.
Added a prototype of two-way sync between angular and boxes models.
Works, but lack of versioning of data means that we can get a loop
where there are conflicting updates in flight at the same time, and
they mirror back and forth infinitely. (e.g. angular sends "a" to
server exactly as boxes sends "b" to client. If they are both received
at the same time, they both get accepted as changes, and send a message
back the other way, repeat ad infinitum). This can be resolved in a
similar way to SVN, by adding a revision to data and refusing to accept
"commits" that conflict with newer versions on the server (in boxes).