Added lastChangeIndex to Box
This allows for querying the change index where the Box was last
changed. This will be 0 if the box has never been changed (still has
value assigned on creation), or >1 if there has been a change to this
box. This looks a little like SVN revisions, and allows easy tracking
of the state of a Box, for example to prevent race conditions updating
a Box over the network.