Rough implementation of versioning
"testString" now has a value and an index, and changes to it are
rejected by the server unless they are based on an up to date index.
This prevents the previous situation where changes could be bounced
back and forth between client and server infinitely, when messages
"crossed over" in flight and kept each other going. This needs to be
neatened up into something that is easier to add into a page for a
number of vars, e.g. a form.