Improved versioning
Angular now sends no update to lift when the change to value is due to
an update FROM lift (this is detected by the index increasing). This
means that there is no "echo" of changes from outside angular. Tested
this by making the edited var be the user's status, and edited value
from two tabs.