Created AjaxDataLinkView, used this in TimesheetView
AjaxDataLinkView encapsulates the logic for keeping a Var[String] in
sync with an Angular model. This still requires the corresponding code
in the Angular controller, will look at how to make this as simple as
possible.
Note that we could also do with a way to allow for multiple fast
updates from the client to Lift without the client needing to receive
back a new change index for each committed change - this should be
possible by permitting repeated commits to a Var[String] from one
AjaxDataLinkView using the same change index, providing that only that
View has been responsible for changing the Var[String] since that
change index. We would then not send anything back to the client from
the View for these accepted changes.