Added header to AjaxView, AjaxDataLinkView now inserts angular tags.
AjaxView now has a renderHeader method, and this is used by
InsertCometView to add a "static" header above the comet it inserts
(this may be empty). This in turn allows for AjaxViews like
AjaxDataLinkView to insert html that will be present as soon as the
page is served, which is useful for inserting an angular directive.
This now allows for a full data link with both the lift and angular
sides working, with just one lift snippet.
Added boxes-data-link directive to angularBoxes module, this provides
the angular side of synchronisation, committing changes made in angular
to the server. Removed code from controllers.js since this is now in
the directive.
Tweaked animate-repeat animations.