Refactored paths, added implicits
Path reactions now use an implicit Txn in apply methods, and have an
alternative "now" method that accepts a shelf and creates a new
transaction.
Added SheetBuilder for transact, and SheetDemo to show use of it.
Removed unused TSwingView.
Added Implicits for transact, currently helps with path reactions by
converting closures to them.