Fixed path implicits
Finally worked out how to do it - can now use Path.now with closures to
avoid having to think about whether it is a path to a plain box, via an
option, or to an option. Also added implicits to go straight from
closures to string, boolean and number views, gives fairly short code.
Unfortunately the "implicit txn: Txn" boiler plate can't be avoided,
but not too bad.