Added unview, new project for transact-based lift
Views can now be unviewed to stop receiving new revisions, Autos have
unauto. This allows for views to be removed "early" before they are
GCed when not referenced.
Added new copy of lift demo, to port to transact. Currently only a
simple angular data source is working.
Added a localShutdown to our CometView, which unviews all views when
the comet actor is no longer connected to anything. This avoids
executing the views while the views and/or actor wait to be GCed, which
should avoid a lot of unnecessary work.