Fixed case objects
We now use some Salat stuff to check whether we are trying to read a
case object, and if so ignore the fields and just return the actual
case object. This is modelled on the code in Grater.fromJSON, which
maybe should be called in fromMap as well, but isn't? Salat may
sensibly assume that if you are using a map, you must have a case class
not a case object...

Also Atom appears to have reformatted a bit, annoyingly.